WeatherListCellをプライベートなウィジェットで再定義
