+++ lib/presentation/pages/detail_page.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:sticky_headers/sticky_headers.dart';
-import 'package:weatherapp_flutter/presentation/pages/weather_list_cell.dart';
 
 class DetailPage extends StatelessWidget {
   final Map<String, List<String>> weather = const {
@@ -51,7 +50,7 @@ class DetailPage extends StatelessWidget {
                     ),
                     content: Column(
                       children: dailyEvents
-                          .map((event) => WeatherListCell())
+                          .map((event) => const _WeatherListCell())
                           .toList(),
                     ),
                   );
@@ -60,6 +59,62 @@ class DetailPage extends StatelessWidget {
             ),
           ],
         ),
+      ),
+    );
+  }
+}
+
+class _WeatherListCell extends StatelessWidget {
+  const _WeatherListCell();
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 20),
+      child: Row(
+        children: [
+          const Text('18:00'),
+          const SizedBox(
+            width: 10,
+          ),
+          SizedBox(
+            width: 50,
+            height: 50,
+            child: Image.asset('assets/images/splash_logo.png'),
+          ),
+          const SizedBox(
+            width: 10,
+          ),
+          const Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                '最高気温:',
+                style: TextStyle(
+                  color: Colors.red,
+                ),
+              ),
+              SizedBox(
+                height: 5,
+              ),
+              Text(
+                '最低気温:',
+                style: TextStyle(
+                  color: Colors.blue,
+                ),
+              ),
+              SizedBox(
+                height: 5,
+              ),
+              Text(
+                '湿度:',
+                style: TextStyle(
+                  color: Colors.green,
+                ),
+              ),
+            ],
+          )
+        ],
       ),
     );
   }
